decoded imagse

+++ modelling/create_job.py
@@ -14,6 +14,8 @@ curr_date=now.strftime("%Y%m%d")
 mem = 36
 gpu_n = 1
 run_time = "5-00:00:00"
+n_jobs = 5
+wait_time = 30
 
 #subj info
 #stim info
@@ -22,6 +24,7 @@ study_dir = f'/user_data/vayzenbe/GitHub_Repos/kornet/modelling'
 
 stim_dir = f'/lab_data/behrmannlab/image_sets/'
 stim_dir =f'/user_data/vayzenbe/image_sets/'
+stim_dir =f'/lab_data/plautlab/imagesets/ecoset'
 
 model_dir = f'/user_data/vayzenbe/GitHub_Repos/vonenet'
 
@@ -84,47 +87,59 @@ def copy_data(train_type):
     return train_dir
 
 
-""" 
-job_name = f'extract_acts'
-print(job_name)
-
-#os.remove(f"{job_name}.sh")
-
-f = open(f"{job_name}.sh", "a")
-f.writelines(setup_sbatch(job_name))
 
 
-f.close()
+model_arch = ['vonecornet_s','cornet_s','voneresnet', 'vit','convnext','resnet50','resnext50','alexnet','vgg19', 'ShapeNet','SayCam']
 
-subprocess.run(['sbatch', f"{job_name}.sh"],check=True, capture_output=True, text=True)
-os.remove(f"{job_name}.sh") """
-
-
-
-for model in model_arch:
-    for train_type in train_types:
-        
-        job_name = f'{model}_{train_type}{suf}_{curr_date}'
+decode_script = False
+if decode_script == True:
+    n_job = 0
+    for model in model_arch:
+        job_name = f'{model}_extract_acts_{curr_date}'
         print(job_name)
-        
+
         #os.remove(f"{job_name}.sh")
-        
+
         f = open(f"{job_name}.sh", "a")
-        script_name = f'python {study_dir}/train.py --data {stim_dir}/{train_type} -o /lab_data/behrmannlab/vlad/kornet/modelling/weights/ --arch {model} --workers 4 -b 128'
+        script_name = f'python {study_dir}/extract_acts.py {model}'
         f.writelines(setup_sbatch(job_name,script_name))
-        
-        
         f.close()
-        
+
         subprocess.run(['sbatch', f"{job_name}.sh"],check=True, capture_output=True, text=True)
-        os.remove(f"{job_name}.sh")
+        os.remove(f"{job_name}.sh") 
+        n_job += 1
+        if n_job == n_jobs:
+            time.sleep(wait_time*60)
+            n_job = 0
+
+model_arch = ['cornet_ff']
+
+train_script = True
+if train_script == True:
+    for model in model_arch:
+        for train_type in train_types:
+            
+            job_name = f'{model}_{train_type}{suf}_{curr_date}'
+            print(job_name)
+            
+            #os.remove(f"{job_name}.sh")
+            
+            f = open(f"{job_name}.sh", "a")
+            script_name = f'python {study_dir}/train.py --data {stim_dir}/{train_type} -o /lab_data/behrmannlab/vlad/kornet/modelling/weights/ --arch {model} --epochs 5 --workers 4 -b 128'
+            f.writelines(setup_sbatch(job_name,script_name))
+            
+            
+            f.close()
+            
+            subprocess.run(['sbatch', f"{job_name}.sh"],check=True, capture_output=True, text=True)
+            os.remove(f"{job_name}.sh")
 
 
-#python {model_dir}/train.py --in_path {stim_dir}/{train_type} -o /lab_data/behrmannlab/vlad/kornet/modelling/weights/ --model_arch {model_arch} --ngpus {gpu_n}
+    #python {model_dir}/train.py --in_path {stim_dir}/{train_type} -o /lab_data/behrmannlab/vlad/kornet/modelling/weights/ --model_arch {model_arch} --ngpus {gpu_n}
 
-#python finetune_models.py --data {stim_dir}/{train_type}/ -o /lab_data/behrmannlab/vlad/kornet/modelling/weights/ --arch {model} -b 64 --workers 4 --epochs 30
+    #python finetune_models.py --data {stim_dir}/{train_type}/ -o /lab_data/behrmannlab/vlad/kornet/modelling/weights/ --arch {model} -b 64 --workers 4 --epochs 30
 
-#python finetune_models.py --data {stim_dir}/{train_type}/ -o /lab_data/behrmannlab/vlad/kornet/modelling/weights/ --arch {model} -b 32 --workers 4 --epochs 30 --resume /lab_data/behrmannlab/vlad/kornet/modelling/weights/{model}_{train_cat}_checkpoint_1.pth.tar 
+    #python finetune_models.py --data {stim_dir}/{train_type}/ -o /lab_data/behrmannlab/vlad/kornet/modelling/weights/ --arch {model} -b 32 --workers 4 --epochs 30 --resume /lab_data/behrmannlab/vlad/kornet/modelling/weights/{model}_{train_cat}_checkpoint_1.pth.tar
